+++ enterprise/frontend/src/embedding-sdk-ee/notifications/index.ts
@@ -1,5 +1,5 @@
-import { PLUGIN_NOTIFICATIONS_SDK } from "embedding-sdk-bundle/components/public/notifications";
 import { EMBEDDING_SDK_CONFIG } from "metabase/embedding-sdk/config";
+import { PLUGIN_NOTIFICATIONS_SDK } from "metabase/plugins";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 
 import { DashboardSubscriptionsButton } from "./DashboardSubscriptionsButton";
