+++ frontend/src/embedding-sdk-bundle/components/public/dashboard/SdkDashboard.tsx
@@ -22,6 +22,7 @@ import {
 import { SdkAdHocQuestion } from "embedding-sdk-bundle/components/private/SdkAdHocQuestion";
 import { useSdkInternalNavigationOptional } from "embedding-sdk-bundle/components/private/SdkInternalNavigation/context";
 import { SdkQuestion } from "embedding-sdk-bundle/components/public/SdkQuestion/SdkQuestion";
+import { DashboardSubscriptionsButton } from "embedding-sdk-bundle/components/public/notifications/DashboardSubscriptionsButton";
 import { useDashboardLoadHandlers } from "embedding-sdk-bundle/hooks/private/use-dashboard-load-handlers";
 import { useExtractResourceIdFromJwtToken } from "embedding-sdk-bundle/hooks/private/use-extract-resource-id-from-jwt-token";
 import { useSdkBreadcrumbs } from "embedding-sdk-bundle/hooks/private/use-sdk-breadcrumb";
@@ -682,10 +683,11 @@ export const SdkDashboard = withPublicComponentWrapper(SdkDashboardInner, {
     | "ParametersList"
     | "FullscreenButton"
     | "ExportAsPdfButton"
-    | "SubscriptionsButton"
     | "InfoButton"
     | "RefreshPeriod"
-  >;
+  > & {
+    SubscriptionsButton: typeof DashboardSubscriptionsButton;
+  };
 
 SdkDashboard.Grid = Dashboard.Grid;
 SdkDashboard.Header = Dashboard.Header;
@@ -694,7 +696,7 @@ SdkDashboard.Tabs = Dashboard.Tabs;
 SdkDashboard.ParametersList = Dashboard.ParametersList;
 SdkDashboard.FullscreenButton = Dashboard.FullscreenButton;
 SdkDashboard.ExportAsPdfButton = Dashboard.ExportAsPdfButton;
-SdkDashboard.SubscriptionsButton = Dashboard.SubscriptionsButton;
+SdkDashboard.SubscriptionsButton = DashboardSubscriptionsButton;
 SdkDashboard.InfoButton = Dashboard.InfoButton;
 SdkDashboard.RefreshPeriod = Dashboard.RefreshPeriod;
