+++ frontend/src/embedding-sdk-bundle/components/public/notifications/DashboardSubscriptionsButton.tsx
@@ -1,4 +1,4 @@
-import { PLUGIN_NOTIFICATIONS_SDK } from "embedding-sdk-bundle/components/public/notifications";
+import { PLUGIN_NOTIFICATIONS_SDK } from "metabase/plugins";
 
 export function DashboardSubscriptionsButton() {
   return <PLUGIN_NOTIFICATIONS_SDK.DashboardSubscriptionsButton />;
