+++ frontend/src/metabase/dashboard/components/DashboardHeader/DashboardHeaderButtonRow/action-buttons.tsx
@@ -1,3 +1,4 @@
+import { PLUGIN_NOTIFICATIONS_SDK } from "metabase/plugins";
 import { Center, Divider } from "metabase/ui";
 
 import { DashboardBookmark } from "../../DashboardBookmark";
@@ -18,7 +19,6 @@ import {
   FullscreenToggle,
 } from "../buttons";
 import { AddLinkOrEmbedButton } from "../buttons/AddLinkOrEmbedButton";
-import { DashboardSubscriptionsButton } from "../buttons/DashboardSubscriptionsButton";
 import { RefreshIndicator } from "../buttons/RefreshIndicator";
 
 import { DashboardSharingMenu } from "./DashboardSharingMenu/DashboardSharingMenu";
@@ -134,7 +134,7 @@ export const dashboardActionButtons: Record<
   // Modular embedding
   [DASHBOARD_ACTION.DASHBOARD_SUBSCRIPTIONS]: {
     enabled: ({ withSubscriptions }) => withSubscriptions,
-    component: () => <DashboardSubscriptionsButton />,
+    component: () => <PLUGIN_NOTIFICATIONS_SDK.DashboardSubscriptionsButton />,
   },
   [DASHBOARD_ACTION.REFRESH_INDICATOR]: {
     enabled: ({ refreshPeriod }) => refreshPeriod != null && refreshPeriod > 0,
